update createStock method to use double for price

+++ src/main/java/com/anhssupercomputer/stocktradingserver/Stock/StockController.java
@@ -40,7 +40,7 @@ public class StockController {
      * @return a message if the creation was successful, an error otherwise
      */
     @PostMapping
-    public String createStock(@RequestBody String name, @RequestBody String ticker, @RequestBody BigDecimal price, @RequestBody int totalVolume) {
+    public String createStock(@RequestBody String name, @RequestBody String ticker, @RequestBody double price, @RequestBody int totalVolume) {
         Stock stock = new Stock(name, ticker, price, totalVolume, service);
         return "Successfully added Stock: \n" + stock.toString();
     }
